refactor: move ProgressWithFeedback to UI root folder
- Relocate ProgressWithFeedback from progress subfolder to the UI components directory
- Update imports in UserDataModal to use the new path
- Update barrel file to export ProgressWithFeedback directly
- Remove unnecessary nested folder structure
